Changed the template functionality
Added a wpas_locate_template() function (may someone just want to get
the path, but not include the file) and changed the functionality of the
wpas_get_template() function (it's more advanced now).

PS: The $priority_child is not needed, because the child theme should
always be priorized (That's the intention of the child theme).