Use single-ticket.php instead of the_content filter
You should use a single-ticket.php instead of the ```the_content```
filter. The problem is, that if you're applying the ```the_content```
filter to a widget, which is loading before the posts content, your
plugins single ticket content will be displayed in the widget area, not
in the posts content area. This is also why for example woocommerce is
solving it with an own template file.