Corrected issue where button text when clicked would display the wrong text
Fixes #240

All that was needed was a check and an output.
To fix this issue I had to check if data-onsubmit was present, and if so
override the default(being the "Submitting..." text).

It should be noted that this issue was present in all places that made
use of wpas_make_button(). So that means the registration form, the
ticket reply form etc.