Fix improper CSS specification for Admin headers
I assume that the purpose of this style tag is to provide formatting for elements with a element matching div.rn-alert. However, the current CSS applies bad rules to all admin headers with the exception of the H1 tag. This is easily fixed by adding the div.rn-alert class to the specifications of all elements.